feat: Update goal creation process to use double for target GPA and implement GoalSheet widget for adding goals

+++ lib/features/goal/data/goal_model.dart
@@ -1,7 +1,7 @@
 class GoalModel {
   String id;
   String name;
-  int targetGpa;
+  double targetGpa;
   bool isAchieved;
   String targetSemesterId;
   String userId;
@@ -29,7 +29,7 @@ class GoalModel {
   GoalModel copyWith({
     String? id,
     String? name,
-    int? targetGpa,
+    double? targetGpa,
     bool? isAchieved,
     String? targetSemesterId,
     String? userId,
